Simplify express server
