+++ package.json
@@ -12,19 +12,16 @@
     "license": "AGPL 3",
     "scripts": {
         "build": "cd app && next build",
-        "start": "npm run build && env NODE_ENV=production node app.js",
-        "dev": "nodemon --ignore app/ app.js",
+        "start": "npm run build && env NODE_ENV=production node server.js",
+        "dev": "nodemon --ignore app/ server.js",
         "style:check": "eslint -c .eslintrc --ext .js,.jsx --ignore-path .gitignore .",
         "style:format": "npm run style:check -- --fix",
         "style:rules": "eslint --print-config .eslintrc"
     },
     "dependencies": {
-        "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
         "@material-ui/core": "^3.9.3",
         "@zeit/next-sass": "^1.0.1",
         "@zeit/next-typescript": "^1.1.1",
-        "babel-plugin-module-resolver": "^3.2.0",
-        "babel-plugin-wrap-in-js": "^1.1.1",
         "body-parser": "^1.19.0",
         "cookie-parser": "^1.4.4",
         "express": "^4.16.4",
@@ -42,6 +39,9 @@
         "react-plotly.js": "^2.3.0"
     },
     "devDependencies": {
+        "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
+        "babel-plugin-module-resolver": "^3.2.0",
+        "babel-plugin-wrap-in-js": "^1.1.1",
         "babel-eslint": "^10.0.1",
         "eslint": "^5.16.0",
         "eslint-config-airbnb": "^17.1.0",
